Revert "わけた (#10)"
This reverts commit e25026a58c062b5ecf2584d1a2139cd9f99994ae.
