+++ Cargo.toml
@@ -1,3 +1,36 @@
-[workspace]
+[package]
+name = "meigenbot-rust"
+version = "0.1.0"
+authors = ["kawaemon <34652535+kawaemon@users.noreply.github.com>"]
+edition = "2018"
 
-members = ["bot_core", "meigen", "interface"]
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0.114", features = ["derive"] }
+serde_yaml = "0.8.13"
+serde_json = "1.0.56"
+
+rand = "0.7.3"
+log = "0.4.8"
+simple_logger = "1.6.0"
+
+mongodb = "1.0.0"
+bson = "1.0.0"
+
+tokio = { version = "0.2.21", features = ["full"] }
+futures = "0.3.5"
+async-trait = "0.1.36"
+readonly = "0.1.6"
+
+warp = "0.2.3"
+percent-encoding = "2.1.0"
+
+[dependencies.serenity]
+git = "https://github.com/serenity-rs/serenity"
+branch = "next"
+default-features = false
+features = ["rustls_backend", "cache", "client", "gateway", "model"]
+
+[profile.release]
+lto = "fat"
