--- bot_core/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "bot_core"
-version = "0.1.0"
-authors = ["kawaemon <34652535+kawaemon@users.noreply.github.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-log = "0.4.8"
-env_logger = "0.7.1"
-
-dotenv = "0.15.0"
-
-anyhow = "1.0.32"
-
-tokio = { version = "0.2.21", features = ["full"] }
-
-async-trait = "0.1.36"
-
-
-meigen = { path = "../meigen" }
-interface = { path = "../interface" }
-
-
-[dependencies.serenity]
-git = "https://github.com/serenity-rs/serenity"
-branch = "next"
-default-features = false
-features = ["rustls_backend", "cache", "client", "gateway", "model"]
-
