--- interface/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "interface"
-version = "0.1.0"
-authors = ["kawak <34652535+kawaemon@users.noreply.github.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-async-trait = "0.1.40"
