--- meigen/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "meigen"
-version = "0.1.0"
-authors = ["kawak <34652535+kawaemon@users.noreply.github.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-thiserror = "1.0.20"
-interface = { path = "../interface" }
-async-trait = "0.1.40"
-tokio = { version = "0.2.22", features = ["full"] }
-serde = { version = "1.0.116", features = ["derive"] }
-serde_yaml = "0.8.13"
-mongodb = "1.1.1"
-log = "0.4.11"
-readonly = "0.1.6"
-rand = "0.7.3"
